* release/1.1.15: (26 commits)
  Update version_check.php
  Update changelog
  README updates
  Add a 'Donate' link to the installer
  Ignore .vs. Not sure why it wasn't ignored before
  Use Tuples to make UpdateDownloader a bit more sane
  Change 'Donate' message on main screen to be less intrusive
  Add Syncthing icon to top right of installer
  Add OxyPlot to third-party components
  Final bit of polish. Get a legend for the colours up
  Fix rounding of max Y axis value
  More tweaking and playing
  More work.
  Playing around with a basic graph
  Put existing tray stuff into a new subfolder
  Take advantage of C# 7 language features
  Fix oldness in Rakefile
  Update to VS2017
  Fix crash when listing conflicts
  Add the --start-syncthing and --stop-syncthing command-line flags
  ...

+++ src/SyncTrayzor/Utils/ShellTools.cs
@@ -12,8 +12,8 @@ namespace SyncTrayzor.Utils
 
         public static Icon GetIcon(string path, bool isFile)
         {
-            var flags = (uint)(SHGFI_ICON | SHGFI_USEFILEATTRIBUTES | SHGFI_LARGEICON);
-            var attribute = isFile ? (uint)FILE_ATTRIBUTE_FILE : (uint)FILE_ATTRIBUTE_DIRECTORY;
+            var flags = SHGFI_ICON | SHGFI_USEFILEATTRIBUTES | SHGFI_LARGEICON;
+            var attribute = isFile ? FILE_ATTRIBUTE_FILE : FILE_ATTRIBUTE_DIRECTORY;
             var shfi = new SHFileInfo();
             IntPtr res;
             try
